pyproject.toml
[project]
name = "langgraph-cli-sample"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.3.25",
    "langchain-openai>=0.3.22",
    "langgraph>=0.4.8",
    "typer>=0.16.0",
    "python-dotenv>=1.0.0",
    "langgraph-checkpoint-sqlite>=2.0.10",
    "aiosqlite>=0.21.0",
    "langgraph-checkpoint-postgres>=2.0.21",
]
